backend/requirements.txt
# Flask and Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.5
python-socketio==5.10.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# AI/ML
google-generativeai==0.3.2

# Authentication & Security
PyJWT==2.8.0
bcrypt==4.1.2
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2
requests==2.31.0

# Development
python-engineio==4.8.0